Refactored config management and mainnet deployment confirmation
