+++ package.json
@@ -30,6 +30,7 @@
     "ethereumjs-util": "^7.0.4",
     "mocha": "^8.1.3",
     "prettier": "^2.1.1",
+    "readline-sync": "^1.4.10",
     "rimraf": "^3.0.2",
     "solc": "0.5.16",
     "thorify": "1.5.3",
