+++ scripts/executeTimelockTransfer.js
@@ -6,8 +6,10 @@ const Timelock = require(config.pathToTimelockJson);
 const GovernorAlpha = require(config.pathToGovernorAlphaJson);
 const V2Factory = require(config.pathToV2FactoryJson);
 const assert = require('assert');
+const readlineSync = require("readline-sync");
 
-let rpcUrl = null;
+let network = null;
+
 if (process.argv.length < 3) 
 {
     console.error("Please specify network, either mainnet or testnet");
@@ -15,16 +17,14 @@ if (process.argv.length < 3)
 } 
 else
 {
-    if (process.argv[2] == "mainnet") rpcUrl = config.mainnetRpcUrl;
-    else if (process.argv[2] == "testnet") rpcUrl = config.testnetRpcUrl;
-    else {
+    network = config.network[process.argv[2]];
+    if (network === undefined) {
         console.error("Invalid network specified");
         process.exit(1);
     }
 }
 
-const web3 = thorify(new Web3(), rpcUrl);
-
+const web3 = thorify(new Web3(), network.rpcUrl);
 web3.eth.accounts.wallet.add(config.privateKey);
 
 executeTimelockTransfer = async() =>
@@ -38,6 +38,12 @@ executeTimelockTransfer = async() =>
     {
         console.log("\n==============================================================================\n");
         console.log("Attempting executeTransaction on Timelock");
+
+        if (network.name == "mainnet")
+        {
+            let input = readlineSync.question("Confirm you want to deploy this on the MAINNET? (y/n) ");
+            if (input != 'y') process.exit(1);
+        }
 
         const timelockContract = new web3.eth.Contract(Timelock.abi, changeAdminConfig.timelockAddress);
